+++ tests/unit/customizations/emr/test_create_default_role.py
@@ -10,12 +10,15 @@
 # distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF
 # ANY KIND, either express or implied. See the License for the specific
 # language governing permissions and limitations under the License.
-
 import mock
-import awscli.customizations.emr.emrutils as emrutils
-import awscli.customizations.emr.createdefaultroles as createdefaultroles
+
 from botocore.compat import json
 from botocore.vendored import requests
+from botocore.exceptions import ClientError
+
+import awscli.customizations.emr.emrutils as emrutils
+import awscli.customizations.emr.createdefaultroles as createdefaultroles
+from awscli.testutils import unittest
 from tests.unit.customizations.emr import EMRBaseAWSCommandParamsTest as \
     BaseAWSCommandParamsTest
 
@@ -142,9 +145,9 @@ class TestCreateDefaultRole(BaseAWSCommandParamsTest):
     @mock.patch('awscli.customizations.emr.emr.'
                 'CreateDefaultRoles._construct_result')
     @mock.patch('awscli.customizations.emr.emr.'
-                'CreateDefaultRoles._check_if_instance_profile_exists')
+                'CreateDefaultRoles.check_if_instance_profile_exists')
     @mock.patch('awscli.customizations.emr.emr.'
-                'CreateDefaultRoles._check_if_role_exists')
+                'CreateDefaultRoles.check_if_role_exists')
     @mock.patch('awscli.customizations.emr.emr.'
                 'CreateDefaultRoles._get_role_policy')
     def test_default_roles_not_exist(self, get_rp_patch,
@@ -161,7 +164,7 @@ class TestCreateDefaultRole(BaseAWSCommandParamsTest):
         self.run_cmd(cmdline, expected_rc=0)
 
         # Only 8 operations will be called as we are mocking
-        # _check_if_role_exists and _check_if_instance_profile_exists methods.
+        # check_if_role_exists and check_if_instance_profile_exists methods.
         self.assertEqual(len(self.operations_called), 6)
 
         self.assertEqual(self.operations_called[0][0].name, 'CreateRole')
@@ -208,9 +211,9 @@ class TestCreateDefaultRole(BaseAWSCommandParamsTest):
     @mock.patch('awscli.customizations.emr.createdefaultroles'
                 '.get_service_principal')
     @mock.patch('awscli.customizations.emr.emr.'
-                'CreateDefaultRoles._check_if_instance_profile_exists')
+                'CreateDefaultRoles.check_if_instance_profile_exists')
     @mock.patch('awscli.customizations.emr.emr.'
-                'CreateDefaultRoles._check_if_role_exists')
+                'CreateDefaultRoles.check_if_role_exists')
     @mock.patch('awscli.customizations.emr.emr.'
                 'CreateDefaultRoles._get_role_policy')
     def test_get_service_principal_parameters(self, get_rp_patch,
@@ -256,14 +259,14 @@ class TestCreateDefaultRole(BaseAWSCommandParamsTest):
     @mock.patch('awscli.customizations.emr.emr.'
                 'CreateDefaultRoles._create_role_with_role_policy')
     @mock.patch('awscli.customizations.emr.emr.'
-                'CreateDefaultRoles._check_if_instance_profile_exists')
+                'CreateDefaultRoles.check_if_instance_profile_exists')
     @mock.patch('awscli.customizations.emr.emr.'
-                'CreateDefaultRoles._check_if_role_exists')
+                'CreateDefaultRoles.check_if_role_exists')
     def test_constructed_result(self, role_exists_patch,
                                 instance_profile_exists_patch,
                                 create_role_patch,
                                 get_role_policy_patch):
-        role_exists_patch.side_effect = side_effect_of_check_if_role_exists
+        role_exists_patch.side_effect = side_effect_ofcheck_if_role_exists
         instance_profile_exists_patch.return_value = False
         create_role_patch.return_value = CREATE_EC2_ROLE_RESULT
         get_role_policy_patch.return_value = EC2_ROLE_POLICY
@@ -295,7 +298,46 @@ class TestCreateDefaultRole(BaseAWSCommandParamsTest):
             EMR_ROLE_ARN)
 
 
-def side_effect_of_check_if_role_exists(*args, **kwargs):
+class TestCreateDefaultRoles(unittest.TestCase):
+
+    def setUp(self):
+        self.session = mock.Mock()
+        self.client = mock.Mock()
+        self.session.create_client.return_value = self.client
+        self.command = createdefaultroles.CreateDefaultRoles(self.session)
+        setattr(self.command, 'iam_endpoint_url', 'https://www.amazonaws.com')
+        self.parsed_globals = mock.Mock()
+        self.parsed_globals.verify_ssl = True
+
+    def testcheck_if_role_exists_raises_client_error(self):
+        error_response = {
+            'Error': {
+                'Code': 400,
+                'Message': 'foo'
+            }
+        }
+        error = ClientError(error_response, 'GetRole')
+        self.client.get_role.side_effect = error
+
+        with self.assertRaises(ClientError):
+            self.command.check_if_role_exists('role', self.parsed_globals)
+
+    def test_check_instance_profile_exists_raises_client_error(self):
+        error_response = {
+            'Error': {
+                'Code': 400,
+                'Message': 'foo'
+            }
+        }
+        error = ClientError(error_response, 'GetInstanceProfile')
+        self.client.get_instance_profile.side_effect = error
+
+        with self.assertRaises(ClientError):
+            self.command.check_if_instance_profile_exists(
+                'role', self.parsed_globals)
+
+
+def side_effect_ofcheck_if_role_exists(*args, **kwargs):
     if args[0] == EC2_ROLE_NAME:
         return False
     else: